+++ app/models/subject.rb
@@ -121,7 +121,7 @@ class Subject < ApplicationRecord
 
   def validate_subject_format
     if user && site && site.subject_regex.present? && site.subject_regex !~ subject_code
-      errors[:base] << "#{project.subject_code_name_full} must be in the following format: #{site.regex_string}"
+      errors[:subject_code] << "must be in the following format: #{site.regex_string}"
     end
   end